student_portfolio/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asgiref=3.4.1=pyhd3eb1b0_0
bzip2=1.0.8=he774522_0
ca-certificates=2022.6.15=h5b45459_0
certifi=2022.6.15=py37h03978a9_0
cffi=1.15.1=py37hd8e9650_0
cryptography=37.0.4=py37h65266a2_0
django=3.2=pyhd3eb1b0_0
django-cors-headers=3.13.0=pyhd8ed1ab_0
djangorestframework=3.13.1=pyhd8ed1ab_0
djongo=1.3.6=pypi_0
dnspython=2.2.1=pyhd8ed1ab_0
idna=3.3=pyhd8ed1ab_0
libffi=3.4.2=hd77b12b_4
libpq=12.9=hb652d5d_3
openssl=1.1.1q=h8ffe710_0
pip=21.2.4=py37haa95532_0
psycopg2=2.8.6=py37hcd4344a_1
pycparser=2.21=pyhd8ed1ab_0
pymongo=3.12.1=pypi_0
python=3.7.13=h6244533_0
python_abi=3.7=2_cp37m
pytz=2022.1=py37haa95532_0
setuptools=61.2.0=py37haa95532_0
sqlite=3.39.2=h2bbff1b_0
sqlparse=0.2.4=py37_0
tk=8.6.12=h2bbff1b_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py37haa95532_2
xz=5.2.5=h8cc25b3_1
zlib=1.2.12=h8cc25b3_2
